Fix some errors

+++ src/main/java/sudokupack/SudokuField.java
@@ -12,7 +12,7 @@ private Integer field;
     }
 
     public int getValue() {
-        Integer value = field;
+        int value = field.intValue();
         return value;
     }
